// libs/enums/pet.enum.ts
export enum PetType {
	DOG = 'DOG',
	CAT = 'CAT',
	BIRD = 'BIRD',
	OTHER = 'OTHER',
}

export enum PetListingType {
	SALE = 'SALE',
	ADOPTION = 'ADOPTION',
}

export enum PetStatus {
	ACTIVE = 'ACTIVE',
	RESERVED = 'RESERVED',
	SOLD = 'SOLD',
	ADOPTED = 'ADOPTED',
	DELETE = 'DELETE',
}

export enum PetGender {
	MALE = 'MALE',
	FEMALE = 'FEMALE',
	UNKNOWN = 'UNKNOWN',
}

export enum PetLocation {
	SEOUL = 'SEOUL',
	BUSAN = 'BUSAN',
	INCHEON = 'INCHEON',
	DAEGU = 'DAEGU',
	GYEONGJU = 'GYEONGJU',
	GWANGJU = 'GWANGJU',
	CHONJU = 'CHONJU',
	DAEJON = 'DAEJON',
	JEJU = 'JEJU',
}